requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.25.2
packaging==24.1
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.2.2
scipy==1.14.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
Werkzeug==3.0.3
openai-whisper==20231117
yt-dlp==2024.4.9
ffmpeg-python==0.2.0
